+++ pytket/docs/changelog.rst
@@ -9,6 +9,7 @@ Features:
 * Update to pytket-circuit-renderer 0.8.
 * Add two new status values for circuits on backends: "CANCELLING" and "RETRYING".
 * Use `lark` package instead of deprecated `lark-parser`.
+* Add ``GreedyPauliSimp`` optimisation pass.
 
 Fixes: